Refactor App; fix issues (hopefully)
