+++ frontend/src/App.module.css
@@ -7,10 +7,6 @@
   display: flex;
 }
 
-.MainView {
-  width: 95%;
-}
-
 .GroupScreen {
   display: flex;
 }
