+++ frontend/src/components/TaskView/ProgressTracker/index.module.css
@@ -14,7 +14,6 @@
 
 .DesktopView {
   width: 70px;
-  height: 85vh;
   margin-right: 1em;
   flex-direction: column-reverse;
 }
