+++ frontend/src/components/TaskView/index.module.css
@@ -2,6 +2,7 @@
   padding: 0 32px;
   display: flex;
   flex-direction: row;
+  height: 85vh;
 }
 
 @media only screen and (min-width: 840px) {
