+++ frontend/src/components/TitleBar/index.module.css
@@ -2,6 +2,7 @@
   position: relative;
   margin: 0;
   padding: 32px 32px 16px 32px;
+  height: 10%;
 }
 
 .Time {
